Track source array names in PrimitiveOperation to fix a bug with argument ordering (#416)
The bug was due to a mismatch between the order of the input source arrays
in the DAG and the function argument order for the operation. Since the DAG
order is not stable, the operation now records the names of the input source
arrays in the source_array_names variable.