Update CUBED_CONFIG path in example README.md
While testing cubed, I was experiencing a lot of new-user environment issues. One issue was that I'd setup `CUBED_CONFIG` to point to a directory rather than a file path, as is shown in this example. However, I was hacking on a problem and had several `yaml` files that I was testing on. Depending on my naming convention, different files might get picked up as my `CUBED_CONFIG`. A better practice is to explicitly name the file in the README example.